feat(load): move load test from dialog to inline tab panel (#62)
